+++ Packages/com.datea83.choyutilities/Runtime/Entities/UI/UIData.cs
@@ -1,0 +1,14 @@
+﻿using Unity.Entities;
+
+namespace EugeneC.ECS {
+	
+	public struct UIData : IComponentData {
+		public byte ParentId;
+		public byte OwnId;
+	}
+
+	[InternalBufferCapacity(1)]
+	public struct UIBuffer : IBufferElementData {
+		public byte Value;
+	}
+}