+++ Packages/com.datea83.choyutilities/Runtime/Entities/UI/UiHandleSystemBase.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using Unity.Entities;
+
+namespace EugeneC.ECS {
+	
+	[UpdateInGroup(typeof(Eu_EffectSystemGroup))]
+	public partial class UiHandleSystemBase : SystemBase {
+
+		public event Action<byte, byte> OnUiEnterHover;
+		public event Action<byte, byte> OnUiExitHover;
+		public event Action<byte, byte> OnUiClicked;
+		
+		protected override void OnCreate() {
+			
+		}
+		
+		protected override void OnUpdate() {
+			
+		}
+	}
+}