+++ Packages/com.datea83.choyutilities/Runtime/Monobehaviour/Generic/GenericSingleton.cs
@@ -1,13 +1,63 @@
 using System;
 using System.Threading;
+using Unity.Entities;
 using UnityEngine;
 
 namespace EugeneC.Singleton {
     
     public abstract class GenericSingleton<T> : MonoBehaviour
         where T : MonoBehaviour {
+	    
         public static T Instance { get; private set; }
+        
+		#region ECS        
+        private const ushort MAX_FRAME = 200;
+        protected World World { get; private set; }
 
+        private async void GetWorld() {
+	        try {
+		        var frame = 0;
+		        while (frame < MAX_FRAME) {
+		        
+			        World = World.DefaultGameObjectInjectionWorld;
+		        
+			        if(World is not null) break;
+			        frame++;
+			        Debug.Log($"{gameObject.name} is waiting for World at frame at {frame}");
+			        await Awaitable.NextFrameAsync(Token);
+		        }
+		        
+		        if (World is null) 
+			        Debug.LogError("World is null, did you forget to add a subscene to the scene?");
+	        }
+	        catch (Exception e) {
+		        Debug.LogError(e);
+	        }
+        }
+
+        protected async Awaitable<TComponent> GetSingletonEntity<TComponent>() 
+	        where TComponent : unmanaged, IComponentData {
+	        
+	        var query = World.EntityManager.CreateEntityQuery(
+		        ComponentType.ReadOnly<TComponent>());
+
+	        TComponent singleton = default;
+	        var validSingleton = false;
+	        var frame = 0;
+	        while (frame < MAX_FRAME) {
+		        validSingleton = query.TryGetSingleton(out singleton);
+		        
+		        if (validSingleton) break;
+		        Debug.Log($"{gameObject.name} is waiting for Singleton at frame at {frame}");
+		        frame++;
+		        await Awaitable.NextFrameAsync(Token);
+	        }
+	        
+	        return !validSingleton ? throw new Exception("Singleton not found") : singleton;
+        }
+		#endregion        
+
+		#region AsyncCancellation
         private CancellationTokenSource _cts = new();
         protected CancellationToken Token => _cts.Token;
         protected event Action OnCancelTask;
@@ -18,6 +68,7 @@ namespace EugeneC.Singleton {
             _cts = new CancellationTokenSource();
             OnCancelTask?.Invoke();
         }
+        #endregion       
 
         protected virtual void InitSingleton() {
             if (Instance is not null && Instance != this) {
@@ -33,11 +84,10 @@ namespace EugeneC.Singleton {
                 Instance = null;
         }
 
-        protected void KeepSingleton(bool keep) {
-            if (keep) DontDestroyOnLoad(this);
+        protected virtual void Awake() {
+	        InitSingleton();
+	        GetWorld();
         }
-
-        protected virtual void Awake() { InitSingleton(); }
 
         protected virtual void OnDisable() { CancelTask(); }